Move VisualizationConfigType to constants
